smart animal

+++ NoseClickerGame/Assets/_Project/_Features/GameUI/Upgrades/UpgrageData.cs
@@ -8,5 +8,5 @@ public class UpgrageData : ScriptableObject
     public UpgradeType Type;
     public string Name;
     public int[] Prices;
-    public int[] Values;
+    public float[] Values;
 }
